mavgen: Add base xml parsing
I tried to follow mavschema.xsd as much as possible (with some
deviations in favor of structure of real definitions). For these
reasons, some of the fields can be unused, but it's okay.

Also, tried to cover as much as possible with tests to be sure that
we can go to the next steps of the code generation.