mavgen: Add the model and normaliser
The model is more like "intermediate representation" of the
flattened module. The parsed xml is normalised to it by simplifying
some types and checking a bunch of stuff, like repeated IDs, names,
etc.

The idea is to have a code generator which accepts this already
normalised module so it can focus solely on code generation.